--- routes/shop.js
@@ -1,35 +1,0 @@
-const fortniteAPI = require('./auth')
-const Store = require('../src/store')
-const Config = require('../src/config')
-
-function getStore (req, res) {
-  let language = req.params.lang || 'en'
-  fortniteAPI.login()
-    .then(() => {
-      fortniteAPI.getStore(language)
-        .then((store) => {
-          Store.convert(store, Config.static_uri)
-            .then((resultStore) => {
-              res.json(resultStore)
-            })
-            .catch((err) => {
-              /* istanbul ignore next */
-              res.status(500).send({
-                code: 500,
-                message: err
-              })
-            })
-        })
-        .catch((err) => {
-          /* istanbul ignore next */
-          res.status(500).send({
-            code: 500,
-            message: err
-          })
-        })
-    })
-}
-
-module.exports = {
-  getStore
-}
